update release.yml to remove unnecessary build targets  (#27)
* feat(release): add GitHub Actions workflow for automated release

This commit adds a new GitHub Actions workflow for automated release. The workflow is triggered on push events with tags that match the pattern `v*`. The workflow builds and releases the Rust project for multiple operating systems and architectures. It also creates a new release and uploads the release artifacts as assets.

* chore(release): update release.yml to remove unnecessary build targets

Removed the build targets for windows and darwin in the release.yml file to streamline the release process and reduce unnecessary build times. Also updated the glob module import to use the correct package name.