feat(git-scripts): add automated commit and PR templates (#31)
Added three new scripts to automate the process of generating commit and PR templates. The commit template script takes the current cached diff and prompts the user to generate a commit message based on a pre-defined template. The PR template script generates a markdown template for creating a new PR based on the changes made in the branch. Also included is a sample script to get chatgpt to do code reviews for rust code. These templates can be definitely improved upon.

This is a new feature that enhances the developer experience and helps maintain consistency in commit and PR messages.